% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceb.R
\name{calc_ceb}
\alias{calc_ceb}
\title{Calculate mean children ever born & living}
\usage{
calc_ceb(
  data,
  by = NULL,
  agegr = 3:10 * 5,
  clusters = ~v021,
  strata = ~v024 + v025,
  weight = "v005",
  ceb = c(ceb = "v201", lc = "v218"),
  dob = "v011",
  intv = "v008",
  varmethod = "lin",
  scale = 12
)
}
\arguments{
\item{data}{[\code{data.frame()}]\cr
Individual recode dataset.}

\item{by}{[\code{formula()}]\cr
A formula specifying factor variables by which to stratify analysis.
Default is 'NULL' to specify no stratification.}

\item{agegr}{[\code{numeric()}]\cr
The start of each age group (in years) to group each individual's age by
and to stratify analysis by. Can be 'NULL' to specify no age-groups should
be created and stratified by. Default is 15, 20, 25, 30, 35, 40, 45, 50.}

\item{clusters}{[\code{formula()}]\cr
A formula specifying cluster ids to be passed as \code{ids} to survey::svydesign.
Must correspond to variables in \code{data}. Default is '~v021'.}

\item{strata}{[\code{formula()}]\cr
A formula specifying strata to be passed as \code{strata} to survey::svydesign.
Must correspond to variables in \code{data}. Default is '~v024+v025'.}

\item{weight}{[\code{character(1)}]\cr
Variable name for sample weight. Default is 'v005'.}

\item{ceb}{[\code{character()}]\cr
Variable names for children ever born related columns to calculate the mean
& standard error of. Value corresponds to name in \code{data}, name corresponds
to new variable name in output. Default is 'c(ceb = "v201", lc = "v218")'.}

\item{dob}{[\code{character(1)}]\cr
Variable name for date of birth of each individual. Default is 'v008'.
Required when \code{agegr} is not 'NULL'.}

\item{intv}{[\code{character(1)}]\cr
Variable name for interview date for each individual. Default is 'v008'.
Required when \code{agegr} is not 'NULL'.}

\item{varmethod}{[\code{character(1)}]\cr
Method for variance calculation. Default is 'lin' for Taylor linearisation.
Other options include 'JKn' for stratified designs & 'JK1' for stratified
designs. See \code{survey::as.svrepdesign} for more information.}

\item{scale}{[\code{numeric(1)}]\cr
Scale for dates inputs (\code{dob}, \code{intv}) to calendar years. 12 for CMC inputs.}
}
\value{
[\code{data.frame()}] consisting of estimates and standard errors as
returned by \code{survey::svyby}. The full covariance matrix of the estimates can
be retreived by \code{vcov(val)}.
}
\description{
Use default arguments to calculate mean children ever born
related statistics from a DHS individual recode file.
}
\examples{
data(zzir)
calc_ceb(zzir)

}
\seealso{
\code{\link[survey:svydesign]{survey::svydesign()}}

survey::svyby()

survey::svymean()
}
