% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling-survival.R
\name{calc_dhs_mx}
\alias{calc_dhs_mx}
\title{Calculate age-specific mortality rates in period preceding survey.
Interval 'period' defined in the months before the survey.
Should replicate mortality rates reported in DHS reports.}
\usage{
calc_dhs_mx(sib, period = c(0, 84))
}
\description{
Calculate age-specific mortality rates in period preceding survey.
Interval 'period' defined in the months before the survey.
Should replicate mortality rates reported in DHS reports.
}
