% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child-mortality.R
\name{calc_nmx}
\alias{calc_nmx}
\title{Calculate the mortality rate between age x and x+n (nmx)}
\usage{
calc_nmx(
  data,
  by = NULL,
  agegr = c(0, 1, 3, 5, 12, 24, 36, 48, 60)/12,
  period = NULL,
  cohort = NULL,
  tips = c(0, 5, 10, 15),
  clusters = ~v021,
  strata = ~v024 + v025,
  weight = "v005",
  dob = "b3",
  dod = "dod",
  death = "death",
  intv = "v008",
  varmethod = "lin",
  origin = 1900,
  scale = 12,
  counts = FALSE,
  clustcounts = FALSE
)
}
\arguments{
\item{data}{A dataset (data.frame), for example a DHS births recode (BR) dataset.}

\item{by}{A formula specifying factor variables by which to stratify analysis.}

\item{agegr}{Numeric vector defining ages \emph{in years} for splits.}

\item{period}{Numeric vector defining calendar periods to stratify analysis, use \code{NULL} for no periods.}

\item{cohort}{Numeric vector defining birth cohorts to stratify analysis, use \code{NULL} for no cohort stratification.}

\item{tips}{Break points for TIme Preceding Survey.}

\item{clusters}{Formula or data frame specifying cluster ids from largest level to smallest level, ‘~0’ or ‘~1’ is a formula for no clusters.}

\item{strata}{Formula or vector specifying strata, use ‘NULL’ for no strata.}

\item{weight}{Formula or vector specifying sampling weights.}

\item{dob}{Variable name for date of birth (character string).}

\item{dod}{Variable name for date of death (character string).}

\item{death}{Variable name for event variable (character string).}

\item{intv}{Variable name for interview date (character string).}

\item{varmethod}{Method for variance calculation. Currently "lin" for Taylor
linearisation or "jk1" for unstratified jackknife, or "jkn", for stratified
jackknife.}

\item{origin}{Origin year for date arguments. 1900 for CMC inputs.}

\item{scale}{Scale for dates inputs to calendar years. 12 for CMC inputs.}

\item{counts}{Whether to include counts of deaths & person-years ('pys')
in the returned `data.frame`. Default is 'FALSE'.}

\item{clustcounts}{Whether to return additional attributes storing cluster
specific counts of deaths `attr(val, 'events_clust')`, person-years
`attr(val, 'pyears_clust')` & number of clusters in each strata
`attr(val, 'strataid')`. Only applicable when using jacknife `varmethod`
'jk1' or 'jkn'. 'strataid' is only included for 'jkn' `varmethod`. Default
is 'FALSE'.}
}
\description{
Default arguments are configured to calculate under 5 mortality rate (nmx)
from a DHS Births Recode file.
}
\examples{

# Calculate age specific nmx from birth history dataset.
data(zzbr)
zzbr$death <- zzbr$b5 == "no"  # b5: child still alive ("yes"/"no")
zzbr$dod <- zzbr$b3 + zzbr$b7 + 0.5
child_mx <- calc_nmx(zzbr)

# Calculate age specific nmx from individual recode.
data(zzir)
zzsib <- reshape_sib_data(zzir)
zzsib$death <- factor(zzsib$mm2, c("dead", "alive")) == "dead"
zzsib$sex <- factor(zzsib$mm1, c("female", "male"))  # drop mm2 = 3: "missing"
adult_mx <- calc_nmx(zzsib, by=~sex, agegr=seq(15, 50, 5), tips=c(0, 7), dob="mm4", dod="mm8")

}
