% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child-mortality-indirect.R
\name{calc_u5mr_indirect}
\alias{calc_u5mr_indirect}
\title{Calculate under-five mortality rates based on indirect demographic techniques}
\usage{
calc_u5mr_indirect(data, by = NULL, clusters = ~v021, strata = ~v024 +
  v025, weight = "v005", intv = "v008", dob = "v011", dfb = "v211",
  ceb = "ceb", csurv = "csurv", origin = 1900, method = "ma",
  family = "Princeton-West", sex = "combined")
}
\arguments{
\item{data}{A dataset (data.frame), for example a DHS Individual recode (IR) dataset.}

\item{by}{A formula specifying factor variables by which to stratify analysis.}

\item{clusters}{Formula or data frame specifying cluster ids from largest level to smallest level, '~0' or '~1' is a formula for no clusters.}

\item{strata}{Formula or vector specifying strata, use 'NULL' for no strata.}

\item{intv}{Variable name for interview date (character string).}

\item{dob}{Variable name for date of birth of the mother (character string).}

\item{dfb}{Variable name for the date of first birth, only needed for using data classified by the time since first birth (character string).}

\item{ceb}{Variable name for the number of children ever born (character string).}

\item{csurv}{Variable name for the number of children surviving (character string).}

\item{origin}{Origin year for date arguments. 1900 for CMC inputs.}

\item{method}{Character string specifying the variant of the method to be used: currently, "ma" for the classical maternal age variant, and "tsfb" for using data classified by the time since first birth.}

\item{family}{Model life table family to use to (1) translate proportion dead into nq0 values and (2) translate age-specific nq0 into U5MR: currently, "West", "North", "South", "East" for Princeton regional model life tables.}

\item{sex}{Character string specifying if sex-specific estimates should be produced ("combined", "males", "females")}

\item{weights}{Formula or vector specifying sampling weights.}
}
\description{
This function estimates under-five mortality rates from information on the mean number of children ever born and children still alive reported by women classified by age group, or classified by time since first birth. Default arguments are configured to calculate under 5 mortality
from a DHS Individual Recode file. Unstratified jackknife is used for variance calculation.v
}
\examples{

data(zzir)

## Calculate 5q0 for both sexes based on the maternal age variant, using DHS
zzir$ceb <- zzir$v201
zzir$csurv <- zzir$v201 - zzir$v206 - zzir$v207
calc_u5mr_indirect(zzir, family = "Princeton-West")

## Calculate 5q0 for males only based on the maternal age variant, using DHS
zzir$ceb_m <- zzir$v202 + zzir$v204 + zzir$v206
zzir$csurv_m <- zzir$ceb_m - zzir$v206
calc_u5mr_indirect(zzir, ceb = "ceb_m", csurv = "csurv_m",
                  family = "Princeton-West", sex = "males")

## Calculate 5q0 for females only based on the maternal age variant, using DHS
zzir$ceb_f <- zzir$v203 + zzir$v205 + zzir$v207
zzir$csurv_f <- zzir$ceb_f - zzir$v207
calc_u5mr_indirect(zzir, ceb = "ceb_f", csurv = "csurv_f",
                  family = "Princeton-West", sex = "females")

## 5q0 by sociodemographic characteristics
calc_u5mr_indirect(zzir, by=~v102) # by urban/rural residence

## indirect 5q0 using time since first birth variant
calc_u5mr_indirect(zzir, method = "tsfb") 

}
\references{
Hill, K. Indirect estimation of child mortality. In: Moultrie, T., et al. (eds.). \emph{Tools for demographic estimation}, 2013, Paris: IUSSP, 148-164, available at \url{http://demographicestimation.iussp.org/content/indirect-estimation-child-mortality}

Verhulst, A. Child mortality estimation: An assessment of summary birth history methods using microsimulation, \emph{Demographic Research}, 2016, 34, 1075-1128
}
