% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling-survival.R
\name{create_tips_data}
\alias{create_tips_data}
\title{Create episode dataset split by period, age group, and time preceding survey indicator (TIPS)}
\usage{
create_tips_data(dat, period = do.call(seq.int, as.list(range(dat$intvy) +
  c(-16, 1))), agegr = 3:12 * 5, tips = 0:15, dobvar = "sibdob",
  dodvar = "sibdod")
}
\description{
Create episode dataset split by period, age group, and time preceding survey indicator (TIPS)
}
