% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_pyears.R
\name{demog_pred_rate}
\alias{demog_pred_rate}
\title{Internal function to predict an estimated rate and standard error}
\usage{
demog_pred_rate(aggr, by, clusters, strata, varmethod, counts, clustcounts)
}
\arguments{
\item{aggr}{output from `demogsurv::demog_pyears`}

\item{by}{A formula specifying factor variables by which to stratify analysis.}

\item{clusters}{Formula or data frame specifying cluster ids from largest level to smallest level, ‘~0’ or ‘~1’ is a formula for no clusters.}

\item{strata}{Formula or vector specifying strata, use ‘NULL’ for no strata.}

\item{varmethod}{Method for variance calculation. Currently "lin" for Taylor
linearisation or "jk1" for unstratified jackknife, or "jkn", for stratified
jackknife.}

\item{counts}{Whether to include counts of deaths & person-years ('pys')
in the returned `data.frame`. Default is 'FALSE'.}

\item{clustcounts}{Whether to return additional attributes storing cluster
specific counts of deaths `attr(val, 'events_clust')`, person-years
`attr(val, 'pyears_clust')` & number of clusters in each strata
`attr(val, 'strataid')`. Only applicable when using jacknife `varmethod`
'jk1' or 'jkn'. 'strataid' is only included for 'jkn' `varmethod`. Default
is 'FALSE'.}
}
\description{
Internal function to predict an estimated rate and standard error
}
