% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_pyears.R
\name{demog_pyears}
\alias{demog_pyears}
\title{Events and person-years from episode data for demographic analysis}
\usage{
demog_pyears(
  formula,
  data,
  period = NULL,
  agegr = NULL,
  cohort = NULL,
  tips = NULL,
  origin = 1900,
  scale = 12,
  dob = "(dob)",
  intv = "(intv)",
  tstart = "tstart",
  tstop = "tstop",
  event = "event",
  weights = NULL
)
}
\description{
This is a wrapper for the \code{\link[survival]{pyears}} function
in the \code{survival} package with convenient stratifications for
demographic analyses.
}
\details{
Note that \code{event} must be a binary variable per the internals
of the \code{\link[survival]{pyears}()} function. The function could
be updated to work around this stipulation.
}
\seealso{
\code{\link[survival]{pyears}}, \code{\link[survival]{tcut}}
}
