% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_rates.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife covariance calculation}
\usage{
jackknife(x, n, strataid = NULL, L = diag(ncol(x)), fn = I)
}
\arguments{
\item{x}{\code{v \times k} matrix specifying weighted numerator for each of
\code{k} PSUs (across columns)}

\item{n}{\code{v \times k} matrix specifying weighted denominator for
each PSU (across columns)}

\item{strataid}{integer or factor vector consisting of id for each strata. Optional,
length should be number of columns of x if supplied.}

\item{L}{\code{q \times v} matrix defining a linear transform}

\item{fn}{function to transorm ratio x/n.}
}
\value{
a data frame with \code{q} rows consisting of estimates calculated as
  \code{fn(L * rowSums(x) / rowSums(n))}, standard error, and 95\% CIs calculated
  on the untransformed scale and then transformed.  The covariance matrix is
  returned as the \code{"var"} attribute and can be accessed by \code{vcov(val)}.
}
\description{
Calculate the covariance matrix for a vector of estimates of the form \code{fn(L * x/n)}
using unstratified (JK1) or stratified (JKn) jackknife calculation removing a
single cluster at a time. The calculation assumes infinite population sampling.
}
\details{
If \code{strataid} is provided, then the stratified (JKn) covariance is calculated, while
if \code{strataid = NULL} then the unstratified (JK1) covariance is calculated. The
latter corresponds to the unstratified jackknife covariance reported in DHS survey
reports. The calculations are equivalent for \code{strataid = rep(1, ncol(x))}.
}
\references{
Pedersen J, Liu J (2012) Child Mortality Estimation: Appropriate
  Time Periods for Child Mortality Estimates from Full Birth Histories.
  PLoS Med 9(8): e1001289. \url{https://doi.org/10.1371/journal.pmed.1001289}.
}
