% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child-mortality-indirect.R
\docType{data}
\name{lts.model}
\alias{lts.model}
\title{Logit of lx values of  model life tables used to convert nq0 into U5MR estimates}
\format{An array with 3 rows and 7 columns (for 7 age groups of mothers) and 9 different families of model life tables
\describe{
  \item{q(1), q(2), q(3), q(5), q(10), q(15), q(20)}{Logits of the lx, the number of survivors at age x, where x is 1,2,3...}
}}
\usage{
lts.model
}
\description{
A dataset containing logits of the lx values of standard life tables for exact ages 1 to 20 to convert age-specific nq0 into U5MR
}
\references{
Moultrie, T., et al. (eds.). \emph{Tools for demographic estimation}, 2013, Paris: IUSSP, 148-164, available at \url{http://demographicestimation.iussp.org/content/indirect-estimation-child-mortality}

Verhulst, A. Child mortality estimation: An assessment of summary birth history methods using microsimulation, \emph{Demographic Research}, 2016, 34, 1075-1128
}
\keyword{datasets}
