% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child-mortality-indirect.R
\docType{data}
\name{multipliers.princeton}
\alias{multipliers.princeton}
\title{Princeton multipliers}
\format{An array with 6 rows and 7 columns (for 7 age groups of mothers) and 4 different families
\describe{
  \item{a(x,j), b(x,j), c(x,j)}{Coefficients used to estimate nq0 from proprtions dead of children ever born}
  \item{e(x,j), f(x,j), g(x,j)}{Coefficients used for estimating the reference periods}
}}
\usage{
multipliers.princeton
}
\description{
A dataset containing coefficients developed for the UN Manual X (1983) for estimating probabilities of dying by exact ages nq0 from proportions dead of children ever born classified by age of mother
}
\references{
Verhulst, A. Child mortality estimation: An assessment of summary birth history methods using microsimulation, \emph{Demographic Research}, 2016, 34, 1075-1128

Hill, K. Indirect estimation of child mortality. In: Moultrie, T., et al. (eds.). \emph{Tools for demographic estimation}, 2013, Paris: IUSSP, 148-164, available at \url{http://demographicestimation.iussp.org/content/indirect-estimation-child-mortality}
}
\keyword{datasets}
