% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mortality.R
\name{reshape_sib_data}
\alias{reshape_sib_data}
\title{Convert respondent-level sibling history data to one row per sibling}
\usage{
reshape_sib_data(data, widevars = grep("^v", names(data), value = TRUE),
  longvars = grep(sibvar_regex, names(data), value = TRUE),
  idvar = "caseid", sib_vars = sub("(.*)_.*", "\\\\1", longvars),
  sib_idvar = "mmidx", sibvar_regex = "^mm[idx0-9]")
}
\arguments{
\item{data}{A dataset as \code{data.frame}.}

\item{widevars}{Character vector of respondent-level variable names to include.}

\item{longvars}{Character vector of variables corresponding to each sibling.}

\item{idvar}{Vector of variable names uniquely identifying each respondent.}

\item{sib_vars}{Vector of same length as longvars giving variable names in long dataset.}

\item{sib_idvar}{Variable name uniquely identifying each sibling record. Should appear amongst \code{sib_vars}.}

\item{sibvar_regex}{Optionally, a regular expression to identify variable names for \code{longvars} from names of \code{data}.}
}
\description{
Convert respondent-level sibling history data to one row per sibling
}
\examples{
data(zzir)

zzsib <- reshape_sib_data(zzir)
zzsib$death <- factor(zzsib$mm2, c("dead", "alive")) == "dead"
zzsib$sex <- factor(zzsib$mm1, c("female", "male"))  # drop mm2 = 3: "missing"
calc_nqx(zzsib, by=~sex, agegr=seq(15, 50, 5), tips=c(0, 7), dob="mm4", dod="mm8")

}
