% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zzbr}
\alias{zzbr}
\title{DHS Model Births Recode Dataset}
\format{
A data frame with 23,666 rows and 105 variables. To keep the model
dataset small only variables starting with 'caseid', 'v0', 'v1', or 'b',
have been included.
}
\source{
\url{https://dhsprogram.com/data/Download-Model-Datasets.cfm}
}
\usage{
zzbr
}
\description{
An example DHS births recode dataset with each row representing a child ever
born to individuals eligible for the women's questionnaire. This data is not
from any actual DHS survey.
}
\keyword{datasets}
