% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zzir}
\alias{zzir}
\title{DHS Model Individual Recode Dataset}
\format{
A data frame with 8,348 rows and 819 variables. To keep the model
dataset small only variables starting with 'caseid', 'v0', 'v1', 'v2', 'b',
or 'mm' have been included.
}
\source{
\url{https://dhsprogram.com/data/Download-Model-Datasets.cfm}
}
\usage{
zzir
}
\description{
An example DHS individual recode dataset with each row representing an
individual eligible for the women's questionnaire. This data is not from any
actual DHS survey.
}
\keyword{datasets}
