% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceb.R
\name{calc_ceb}
\alias{calc_ceb}
\title{Calculate mean children ever born (ceb)}
\usage{
calc_ceb(
  data,
  by = NULL,
  agegr = 3:10 * 5,
  clusters = ~v021,
  strata = ~v024 + v025,
  weight = "v005",
  awfactt = NULL,
  ceb = "v201",
  dob = "v011",
  intv = "v008",
  varmethod = "lin",
  scale = 12,
  awfactt_scale = 100,
  counts = FALSE,
  clustcounts = FALSE
)
}
\arguments{
\item{data}{[\code{data.frame()}]\cr
Individual recode dataset.}

\item{by}{[\code{formula()}]\cr
A formula specifying factor variables by which to stratify analysis.
Default is 'NULL' to specify no stratification. Passed to \code{\link[survey:svyby]{survey::svyby()}}.}

\item{agegr}{[\code{numeric()}]\cr
The start of each age group (in years) to group each individual's age by
and to stratify analysis by. Can be 'NULL' to specify no age-groups should
be created and stratified by. Default is 15, 20, 25, 30, 35, 40, 45, 50.}

\item{clusters}{[\code{formula()}]\cr
A formula specifying cluster ids to be passed as \code{ids} to survey::svydesign.
Must correspond to variables in \code{data}. Default is '~v021'.}

\item{strata}{[\code{formula()}]\cr
A formula specifying strata to be passed as \code{strata} to survey::svydesign.
Must correspond to variables in \code{data}. Default is '~v024+v025'.}

\item{weight}{[\code{character(1)}]\cr
Variable name for sample weight. Default is 'v005'.}

\item{ceb}{[\code{character(1)}]\cr
Variable name for children ever born related variable to calculate the mean
& standard error of. Default is 'v201'.}

\item{dob}{[\code{character(1)}]\cr
Variable name for date of birth of each interviewee. Default is 'v008'.
Required when \code{agegr} is not 'NULL'.}

\item{intv}{[\code{character(1)}]\cr
Variable name for interview date of each interviewee. Default is 'v008'.
Required when \code{agegr} is not 'NULL'.}

\item{varmethod}{[\code{character(1)}]\cr
Method for variance calculation. Default is 'lin' for Taylor linearisation.
Other options include 'JKn' for stratified designs & 'JK1' for stratified
designs. See \code{survey::as.svrepdesign} for more information.}

\item{scale}{[\code{numeric(1)}]\cr
Scale for dates inputs (\code{dob}, \code{intv}) to calendar years. 12 for CMC inputs.}
}
\value{
[\code{data.frame()}] consisting of estimates and standard errors.
The full covariance matrix of the estimates can be retrieved by \code{vcov(val)}.
}
\description{
Calculate mean children ever born or other related statistics
(like mean number of living children) from a DHS individual recode file.
}
\examples{
data(zzir)

# children ever born
calc_ceb(zzir)
# children ever born and living
calc_ceb(zzir, ceb = "v218")

}
\seealso{
\code{\link[survey:svydesign]{survey::svydesign()}}, \link[survey:as.svrepdesign]{survey::as.svrepdesign}, \code{\link[survey:svyby]{survey::svyby()}},
& \code{\link[survey:surveysummary]{survey::svymean()}} which this function uses internally.
}
