% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling-survival.R
\name{create_tips_data}
\alias{create_tips_data}
\title{Create episode dataset split by period, age group, and time preceding survey indicator (TIPS)}
\usage{
create_tips_data(
  dat,
  period = do.call(seq.int, as.list(range(dat$intvy) + c(-16, 1))),
  agegr = 3:12 * 5,
  tips = 0:15,
  dobvar = "sibdob",
  dodvar = "sibdod"
)
}
\arguments{
\item{dat}{A dataset as `data.frame`.}

\item{period}{Numeric vector defining calendar periods to stratify analysis, use `NULL` for no periods.}

\item{agegr}{Numeric vector defining ages *in years* for splits.}

\item{tips}{Break points for TIme Preceding Survey.}

\item{dobvar}{Variable name for date of birth (character string).}

\item{dodvar}{Variable name for date of death (character string).}
}
\description{
Create episode dataset split by period, age group, and time preceding survey indicator (TIPS)
}
