% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_rates.R
\name{.mm_aggr}
\alias{.mm_aggr}
\title{Construct a model matrix for aggregating over age groups}
\usage{
.mm_aggr(mf, agegr)
}
\arguments{
\item{mf}{Model frame for predicted rates}

\item{agegr}{Numeric vector defining ages \emph{in years} for splits.}
}
\description{
Construct a model matrix for aggregating over age groups
}
